Move CustomScheduler to utils module
