+++ pyresample/test/test_bucket.py
@@ -11,25 +11,7 @@ except ImportError:
 
 from pyresample.geometry import AreaDefinition
 import pyresample.bucket as bucket
-
-
-class CustomScheduler(object):
-    """Custom dask scheduler that raises an exception if dask is computed
-    too many times."""
-
-    def __init__(self, max_computes=1):
-        """Set starting and maximum compute counts."""
-        self.max_computes = max_computes
-        self.total_computes = 0
-
-    def __call__(self, dsk, keys, **kwargs):
-        """Compute dask task and keep track of number of times we do so."""
-        import dask
-        self.total_computes += 1
-        if self.total_computes > self.max_computes:
-            raise RuntimeError("Too many dask computations were scheduled: "
-                               "{}".format(self.total_computes))
-        return dask.get(dsk, keys, **kwargs)
+from pyresample.test.utils import CustomScheduler
 
 
 class Test(unittest.TestCase):
