+++ pyresample/test/utils.py
@@ -178,3 +178,22 @@ def create_test_latitude(start, stop, shape, twist_factor=0.0, dtype=np.float32)
     lat_array = np.repeat(lat_col, shape[1], axis=1)
     lat_array += twist_array
     return lat_array
+
+
+class CustomScheduler(object):
+    """Custom dask scheduler that raises an exception if dask is computed
+    too many times."""
+
+    def __init__(self, max_computes=1):
+        """Set starting and maximum compute counts."""
+        self.max_computes = max_computes
+        self.total_computes = 0
+
+    def __call__(self, dsk, keys, **kwargs):
+        """Compute dask task and keep track of number of times we do so."""
+        import dask
+        self.total_computes += 1
+        if self.total_computes > self.max_computes:
+            raise RuntimeError("Too many dask computations were scheduled: "
+                               "{}".format(self.total_computes))
+        return dask.get(dsk, keys, **kwargs)
